refactor: Convert il2p Direwolf tests to direct Go test functions
Move rec_count, text, and polarity vars to il2p_test_util.go since that
non-test file references them in multi_modem_process_rec_packet_fake.
