+++ src/il2p_direwolf_test.go
@@ -17,9 +17,7 @@ import (
  *
  *--------------------------------------------------------------*/
 
-func il2p_test_main(t *testing.T) {
-	t.Helper()
-
+func Test_IL2P(t *testing.T) {
 	IL2P_TEST = true
 
 	var enable_color = 1
@@ -745,20 +743,6 @@ func all_frame_types(t *testing.T) {
 
 var addrs2 = "AA1AAA-1>ZZ9ZZZ-9"
 var addrs3 = "AA1AAA-1>ZZ9ZZZ-9,DIGI*"
-
-var text = `'... As I was saying, that seems to be done right - though I haven't time to look it over thoroughly just now - and that shows that there are three hundred and sixty-four days when you might get un-birthday presents -'
-'Certainly,' said Alice.
-'And only one for birthday presents, you know. There's glory for you!'
-'I don't know what you mean by \"glory\",' Alice said.
-Humpty Dumpty smiled contemptuously. 'Of course you don't - till I tell you. I meant \"there's a nice knock-down argument for you!\"'
-'But \"glory\" doesn't mean \"a nice knock-down argument\",' Alice objected.
-'When I use a word,' Humpty Dumpty said, in rather a scornful tone, 'it means just what I choose it to mean - neither more nor less.'
-'The question is,' said Alice, 'whether you can make words mean so many different things.'
-'The question is,' said Humpty Dumpty, 'which is to be master - that's all.'
-`
-
-var rec_count = -1 // disable deserialized packet test.
-var polarity = 0
 
 func test_serdes(t *testing.T) {
 	t.Helper()
