--- src/il2p_test.go
@@ -1,7 +1,0 @@
-package direwolf
-
-import "testing"
-
-func Test_IL2P(t *testing.T) {
-	il2p_test_main(t)
-}
